Add NewTodoCard component and related styles Post function
The commit introduces the NewTodoCard component along with the corresponding CSS styling. The newly added component allows users to enter a new task. Minor adjustments were also made to existing styles and components, such as the TodoCard and TodoColumn, to accommodate this addition. The NewTodoCard component has been configured to only show up in the "OPEN" status section.