Add functionality to move todo item to the right in TODO Application
This change introduces the capability for todo items to be moved across stages within the app. A "Move" button has been added to each todo item unless the status is 'DONE'. Clicking the move button will let users transition items from 'OPEN' to 'IN_PROGRESS' or from 'IN_PROGRESS' to 'DONE'. Additionally, a 'key' prop was added in 'TodoColumn' for better list rendering and a new CSS class was introduced for the 'move-button'.